some updates in DB and doctor part

+++ Medic/src/java/com/medic/facade/AbstractFacade.java
@@ -3,44 +3,52 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-
 package com.medic.facade;
 
 import java.util.List;
 import javax.annotation.Resource;
 import javax.persistence.EntityManager;
 
-
 /**
  *
  * @author Rimma
  */
 public abstract class AbstractFacade<T> {
+
     private Class<T> entityClass;
 
     public AbstractFacade(Class<T> entityClass) {
-        this.entityClass = entityClass;
+        if (entityClass != null) {
+            this.entityClass = entityClass;
+        }
     }
 
     protected abstract EntityManager getEntityManager();
 
     public void create(T entity) {
-        getEntityManager().persist(entity);
-        getEntityManager().flush();
-        
+        if (entity != null) {
+            getEntityManager().persist(entity);
+            getEntityManager().flush();
+        }
     }
 
     public void edit(T entity) {
-        getEntityManager().merge(entity);
+        if (entity != null) {
+            getEntityManager().merge(entity);
+        }
     }
 
     public void remove(T entity) {
-        getEntityManager().remove(getEntityManager().merge(entity));
+        if (entity != null) {
+            getEntityManager().remove(getEntityManager().merge(entity));
+        }
     }
 
     public T find(Object id) {
-        return getEntityManager().find(entityClass, id);
-        
+        if (id != null) {
+            return getEntityManager().find(entityClass, id);
+        }
+        else return null;
     }
 
     public List<T> findAll() {
@@ -49,5 +57,4 @@ public abstract class AbstractFacade<T> {
         return getEntityManager().createQuery(cq).getResultList();
     }
 
-    
 }
